Trait merino_cache::domain::CacheKey

pub trait CacheKey<'a> {
    fn cache_key(&self) -> Cow<'a, str>;
}

An object that can generate a cache key for itself.

Required methods

fn cache_key(&self) -> Cow<'a, str>

Generate a cache key for this object. Two objects that have the same cache key should be functionally identical.

Cache keys should make it clear that they are cache keys, and specify the type of object they refer to. They should also include a version indicator. For example: cache:req:v1:d1bc8d3ba4afc7e1. Excessively long key lengths should be avoided. 100 bytes is a good upper bound.

Loading content...