Struct merino_cache::memory::LOCK_TABLE

struct LOCK_TABLE {
    __private_field: (),
}

Fields
__private_field: ()
Implementations

impl LOCK_TABLE

fn is_locked(&self, key: &str) -> bool

Check to see if there’s any lock for a given key

fn add_lock(&self, key: &str, lock_timeout: Duration) -> Instant

Generate a lock for the given key and timeout

fn update<F>(&self, key: &str, lock: Instant, func: F) where
F: FnMut(),

run func and remove lock, only if the lock we have matches what is registered for the key.

fn prune(&self, start: &Instant)

remove any expired elements from the Pending table (There shouldn’t be many.)