Module merino_cache::redis::domain

Wrapper types for [merino-suggest_traits] types so traits can be implemented

Structs

RedisSuggestions

A wrapper around a Vec of Suggestions that can be inserted into and retrieved from a Redis DB.

Enums

RedisTtl

The result from the Redis TTL command, converting the two error codes (-1 and -2) into enum variants.

Constants

SERIALIZATION_VERSION

A byte string prepended to all cached data to avoid format problems. Cache entries with a different cache key will be treated as invalid.