Struct merino_cache::redis::domain::RedisSuggestions

pub(crate) struct RedisSuggestions(pub(crate) Vec<Suggestion>);

A wrapper around a Vec of Suggestions that can be inserted into and retrieved from a Redis DB.