Struct merino_cache::RedisSuggester

pub struct RedisSuggester<S> where
    S: SuggestionProvider<'a>,  {
    inner: S,
    redis_connection: ConnectionManager,
    default_ttl: Duration,
}

A suggester that uses Redis to cache previous results.

Fields
inner: S

The suggester to query on cache-miss.

redis_connection: ConnectionManager

Connection to Redis.

default_ttl: Duration

The default amount of time a cache entry is valid, unless overridden by inner.

Implementations

impl<S> Suggester<S> where
S: SuggestionProvider<'a>,

pub async fn new_boxed(
settings: &Settings,
provider: S
) -> Result<Box<Self>, SetupError>

Create a WikiFruit provider from settings.

async fn get_key(&self, key: &str) -> Result<CacheCheckResult, SuggestError>

Retrieve an item from the cache

If the item retrieved cannot be deserialized, it will be deleted. If there is no TTL for the retrieved item, one will be added to it.

fn queue_store_key(
&self,
key: &str,
suggestions: Vec<Suggestion>
) -> Result<(), SuggestError>

Queue a command to store an entry in the cache.

This runs as a separate task, and this function returns before the operation is complete.

Errors

Returns an error if the command cannot be queued. Does not error if the command fails to run to completion.

fn queue_delete_key(&self, key: &str) -> Result<(), SuggestError>

Queue a command to delete a key from the cache.

This runs as a separate task, and this function returns before the deletion is complete.

Errors

Returns an error if the command cannot be queued. Does not error if the command fails to run to completion.

fn queue_set_key_ttl(
&self,
key: &str,
ttl: Duration
) -> Result<(), SuggestError>

Queue a command to set the TTL of a key in the cache.

This runs as a separate task, and this function returns before the operation is complete.

Errors

Returns an error if the command cannot be queued. Does not error if the command fails to run to completion.